requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
torch
torchvision
pillow
kaggle
kagglehub
tqdm
flask
flask-cors
